Match contributions by comparing specific names, not entire strings.
The idea here is that for example "Agata Beata" should be able to match
against "Beata A".

To achieve this, name strings are split into lists of specific names, with
which a specific name from the other contribution may or may not be matched.
The matching between specific names in turn then considers the possibility
of it being an initial or a double-initial. Miss-spellings are countered by
comparing using edit distance instead of equivalence.
Since names are checked one by one, the order of the names also
no longer matter, and it is no longer necessary for two string to have the
exact same number of words/names ("Agata Beata" will match against
"Agata Beata Cindy").